added test class for box persistence
